2 Calculadoras criadas
Calculadoras para retângulo e trapézio criadas

+++ xGeometry/xGeometry/XGMEsfera.m
@@ -30,7 +30,7 @@
     [str appendString:[NSString stringWithFormat:@"Considerando π como 3,14\nVe = 4/3 . π . r³\n"]];
     [str appendString:[NSString stringWithFormat:@"Ve = 4/3 . π . %d³\n",self.raio]];
     [str appendString:[NSString stringWithFormat:@"Ve = 4/3 . %d . π\n",self.raio * self.raio * self.raio]];
-    [str appendString:[NSString stringWithFormat:@"Ve = %d . π\n",(self.raio * self.raio * self.raio*4)/3]];
+    [str appendString:[NSString stringWithFormat:@"Ve = %f . π\n",(float)(self.raio * self.raio * self.raio*4)/3]];
     [str appendString:[NSString stringWithFormat:@"Ve = %f",((self.raio * self.raio * self.raio*4)/3) * 3.14]];
     
     return str;
